+++ lib/index.ts
@@ -42,9 +42,6 @@
  *   createProtopediaApiCustomClient
  * } from '@f88/promidas/fetcher';
  *
- * // Note: fetchAndNormalizePrototypes is deprecated.
- * // Use ProtopediaApiCustomClient.fetchPrototypes instead.
- *
  * // In-memory store
  * import { PrototypeInMemoryStore } from '@f88/promidas/store';
  *